refactor utils for json fields
